fix-bug: fix the type of wta to float instead of int

+++ ProcessStructs/PCB.h
@@ -15,7 +15,7 @@ struct PCB
     int remainingTime;
     int waitingTime;
     int turnaroundTime;
-    int weightedTurnaroundTime;
+    float weightedTurnaroundTime;
     
 };
